Added class: GameDtoConverter and ShipDTO.
Instead of list of Ships and a list of all Coordinates for each player, each player now have a list of activeShips and sunkenShips with list of Coordinates inside each Ship specific to that Ship.

Replaced Row and Column in Strike to Coordinate.

In GameEvent: removed strikeRow & Column and isHit, frontend will have to use data from last Strike for its message log.